requirements.txt
sentence-transformers==2.2.2
torch>=2.0.0
faiss-cpu==1.7.4
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
sqlalchemy>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
networkx>=3.0